m8: use custom ril class to fix sprint 'no-sim' bug
Also fixes stock mms app not allowing sending of messages.
This opens the door for all future custom ril class mods
for m8 as it is set for all models. Sprint hack only applys
to sprint. Function that is overridden is pulled from
opt/telephony (no class before so it was using this one origionally).

Signed-off-by: Simon Sickle <simon@simonsickle.com>